+++ src/test/java/io/github/bartmy/http/HelloServiceTest.java
@@ -43,7 +43,7 @@ public class HelloServiceTest {
         //given
         var mockRepository = new LangRepository(){
             @Override
-            public Optional<Lang> findById(Long id) {
+            public Optional<Lang> findById(Integer id) {
                 return Optional.empty();
             }
         };
@@ -83,7 +83,7 @@ public class HelloServiceTest {
     private LangRepository fallbackIdLangRepository() {
         return new LangRepository(){
             @Override
-            public Optional<Lang> findById(Long id) {
+            public Optional<Lang> findById(Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())){
                     return Optional.of(new Lang(null,FALLBACK_ID_WELCOME, null));
                 }
@@ -95,7 +95,7 @@ public class HelloServiceTest {
     private LangRepository alwaysReturningHelloRepository() {
         return new LangRepository(){
             @Override
-            public Optional<Lang> findById(Long id) {
+            public Optional<Lang> findById(Integer id) {
                 return Optional.of(new Lang(null, WELCOME, null));
             }
         };
